requirements.txt
python-daemon
tensorboardX
torchvision
matplotlib
torch
numpy
